Addresses several issues at once. Rotations is now applied to shapes  before drawing happens (Issue 765). All shapes now compute an Envelope3D rather than a stock JTS Envelope (Issue 763). AffineTransformations are now done in 3D by default (Issue 762). Agents are correctly selected in torus mode (Issue 745). Less copies of shapes are made when displaying or transforming them. Atrributes of shapes are now preserved when they are transformed.
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@8395 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.jogl/src/msi/gama/jogl/scene/ImageObject.java
@@ -19,14 +19,14 @@ public class ImageObject extends AbstractObject {
 	public double z;
 	public double width;
 	public double height;
-	public Integer angle = 0;
+	public Double angle = 0d;
 	public boolean isDynamic;
 	public MyTexture texture;
 	public String name;
 
 	public ImageObject(final BufferedImage image, final IAgent agent, final double z_layer, final int layerId,
 		final double x, final double y, final double z, final Double alpha, final double width, final double height,
-		final Integer angle, final GamaPoint offset, final GamaPoint scale, final boolean isDynamic,
+		final Double angle, final GamaPoint offset, final GamaPoint scale, final boolean isDynamic,
 		final MyTexture texture, final String name) {
 		super(null, offset, scale, alpha);
 		setZ_fighting_id((double) layerId);
